refactor: Use simplier file convention
