+++ examples/react-router-7-app/app/root.tsx
@@ -1,4 +1,4 @@
-import "./index.css";
+import "../index.css";
 import { Outlet, Scripts, ScrollRestoration } from "react-router";
 import Providers from "../providers";
