+++ examples/react-router-7-app/providers.tsx
@@ -1,6 +1,6 @@
+import "./fetchClient";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import "./fetchClient";
 import { queryClient } from "./queryClient";
 
 export default function Providers({ children }: { children: React.ReactNode }) {
